requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt[crypto]>=5.3
django-filter>=24.2
drf-spectacular>=0.27
Pillow>=10.4
celery>=5.4
redis>=5.0